% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_TP_CN_mat.R
\name{build_TP_CN_mat}
\alias{build_TP_CN_mat}
\title{create copy number matrix for transition points}
\usage{
build_TP_CN_mat(list_TP_CN)
}
\arguments{
\item{list_TP_CN}{list of data frame with TP position in the first column and the copy number of the TP in the second column}
}
\value{
a matrix
}
\description{
create copy number matrix for transition points
}
\examples{
df1 <- data.frame(TP = c("TP1", "TP2", "TP3", "TP4"), CN = c(1, 2, 3, 4))
df2 <- data.frame(TP = c("TP2", "TP3", "TP4"), CN = c(1, 5, 6))
df3 <- data.frame(TP = c("TP1", "TP5", "TP3", "TP7"), CN = c(11, 22, 35, 14))
df_list <- list(df1, df2, df3)
names(df_list) <- c("df1", "df2", "df3")
build_TP_CN_mat(df_list)

}
