% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_positional_CN.R
\name{get_positional_CN}
\alias{get_positional_CN}
\title{get copy number by provided position}
\usage{
get_positional_CN(BAM_path, pattern, bin_size, copyNumbersCalled_obj, chr, pos)
}
\arguments{
\item{BAM_path}{path of the .bam files}

\item{pattern}{pattern for the project samples}

\item{bin_size}{bin size to calculate copy number by QDNAseq}

\item{copyNumbersCalled_obj}{a QDNAseq object}

\item{chr}{chromosome: single or multiple in a vector}

\item{pos}{position (in bp): single or multiple in a vector (according to the chromosome)}
}
\value{
a data frame
}
\description{
get copy number by provided position
}
\examples{
BAM_path <- "~/Documents/Research/prostate_cancer/PEACE/samples/tumour_BAM_subset/"
bin_size <- 500
pattern <- "PEA310"
copyNumbersCalled_obj <- generate_copyNumberCalled_obj(BAM_path, pattern, bin_size)
chr <- c(1, 2, 13)
pos <- c(1000301, 14000000, 60000000)
get_positional_CN(BAM_path, pattern, bin_size, copyNumbersCalled_obj, chr, pos)

}
