% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumour_samples.R
\name{tumour_samples}
\alias{tumour_samples}
\title{tumour samples}
\usage{
tumour_samples(BAM_path, pattern)
}
\arguments{
\item{BAM_path}{character string for BAM directory}

\item{pattern}{unique pattern tumour samples}
}
\value{
character vector
}
\description{
tumour samples
}
\examples{
pattern <- c(paste0("PEA310", "_1"))
BAM_path <- "~/Documents/Research/prostate_cancer/PEACE/samples/tumour_BAM_subset/"
tumour_samples(BAM_path, pattern)
}
