% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_TPs.R
\name{generate_TP_binary_matrix}
\alias{generate_TP_binary_matrix}
\title{create Transition Point binary matrics}
\usage{
generate_TP_binary_matrix(BAM_path, pattern, copyNumbersCalled_obj)
}
\arguments{
\item{BAM_path}{path of the .bam files}

\item{pattern}{pattern for the project samples}

\item{copyNumbersCalled_obj}{a QDNAseq object}

\item{bin_size}{bin size to calculate copy number by QDNAseq}
}
\value{
a binary matrix
}
\description{
create Transition Point binary matrics
}
\examples{
BAM_path <- "~/Documents/Research/prostate_cancer/PEACE/samples/tumour_BAM_subset/"
bin_size <- 500
pattern <- c(paste0("PEA310", "_1"))
seed = 101
copyNumbersCalled_obj <- generate_copyNumberCalled_obj(BAM_path, pattern, bin_size, seed)
generate_TP_binary_matrix(BAM_path, pattern, copyNumbersCalled_obj)

}
