% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{\%>\%}
\alias{\%>\%}
\alias{generate_CN_df}
\title{generate CN df}
\usage{
generate_CN_df(BAM_path, pattern_PEA, bin_size)
}
\arguments{
\item{BAM_path}{path of .bam files}

\item{pattern_PEA}{pattern for the project samples}

\item{bin_size}{bin size to calculate copy number by QDNAseq}
}
\value{
a dataframe
}
\description{
generate CN df
}
\examples{
BAM_path <- "~/Documents/Research/prostate_cancer/PEACE/samples/tumour_BAM/"
bin_size=500
pattern_PEA = "PEA310"
generate_CN_df(BAM_path,pattern_PEA, bin_size)


'\%>\%' <- magrittr::"\%>\%"
}
