% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_copyNumbersCalled_obj.R
\name{\%>\%}
\alias{\%>\%}
\alias{generate_copyNumberCalled_obj}
\title{generate copy number call object}
\usage{
generate_copyNumberCalled_obj(BAM_path, pattern, bin_size)
}
\arguments{
\item{BAM_path}{path of .bam files}

\item{pattern}{pattern for the project samples}

\item{bin_size}{bin size to calculate copy number by QDNAseq}
}
\value{
a QDNAseq object
}
\description{
generate copy number call object
}
\examples{
BAM_path <- "~/Documents/Research/prostate_cancer/PEACE/samples/tumour_BAM_subset/"
bin_size=500
pattern = "PEA310"
generate_copyNumberCalled_obj(BAM_path,pattern, bin_size)



}
